requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
stripe==6.7.0
supabase

# AI Guardrails Dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# Platform Integrations
openpyxl>=3.1.0


# Session Management
Flask-Session==0.5.0
redis==5.0.1

# PostgreSQL Database
psycopg2-binary==2.9.9



# OAuth Authentication
Authlib==1.3.0

